:recycle: api: move back getByScopeAndReconciliationDate to its context
