+++ api/src/certification/shared/infrastructure/repositories/versions-repository.js
@@ -3,7 +3,6 @@
  */
 
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
-import { NotFoundError } from '../../../../shared/domain/errors.js';
 import { Version } from '../../domain/models/Version.js';
 
 /**
@@ -31,41 +30,6 @@ export const findLatestByScope = async ({ scope }) => {
 
   if (!versionData) {
     return null;
-  }
-
-  return _toDomain(versionData);
-};
-
-/**
- * @param {Object} params
- * @param {Frameworks} params.scope
- * @param {Date} params.reconciliationDate
- * @returns {Promise<Version>}
- */
-export const getByScopeAndReconciliationDate = async ({ scope, reconciliationDate }) => {
-  const knexConn = DomainTransaction.getConnection();
-
-  const versionData = await knexConn('certification_versions')
-    .select(
-      'id',
-      'scope',
-      'startDate',
-      'expirationDate',
-      'assessmentDuration',
-      'globalScoringConfiguration',
-      'competencesScoringConfiguration',
-      'challengesConfiguration',
-    )
-    .where({ scope })
-    .andWhere('startDate', '<=', reconciliationDate)
-    .andWhere((queryBuilder) => {
-      queryBuilder.whereNull('expirationDate').orWhere('expirationDate', '>', reconciliationDate);
-    })
-    .orderBy('startDate', 'desc')
-    .first();
-
-  if (!versionData) {
-    throw new NotFoundError('No certification framework version found for the given scope and reconciliationDate');
   }
 
   return _toDomain(versionData);
